requirements.txt
Django==4.1.6
django-oauth-toolkit==2.2.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
jsonschema==4.17.3
psycopg2-binary==2.9.5
gunicorn==20.1.0
python-dotenv==0.21.1
requests==2.28.2
wheel==0.36.2
httplib2==0.21.0
